+++ esp32/sdkconfig.defaults.esp32p4-eth-wifi
@@ -10,9 +10,12 @@
 # to the P4 as the standard esp_wifi_* API by the esp_wifi_remote + esp_hosted
 # managed components (pulled P4-only via the `rules` gate in main/idf_component.yml).
 # This is a DELIBERATE v6.0-floor exception (see docs/building.md § ESP-IDF version).
-# The bring-up prelude (esp_hosted_init + connect-to-slave, before esp_wifi_init)
-# lives in ensureWifiInit() in src/platform/esp32/platform_esp32.cpp, gated on
-# platform::usesRemoteWifi.
+# No bring-up code is needed in the firmware: esp_hosted self-initialises at boot
+# via a constructor and connects to the C6 on its own. ensureWifiInit() in
+# src/platform/esp32/platform_esp32.cpp explicitly does NOT call esp_hosted_init /
+# connect_to_slave (an earlier prelude that did was removed: it reset the slave and
+# tore down the live SDIO link, bench-proven 2026-06-12). The co-processor firmware
+# read-out is gated on platform::hasWifiCoprocessor.
 #
 # Values below mirror Espressif's own P4 reference fragment:
 #   managed_components/espressif__esp_hosted/examples/host_network_split__power_save/
